@@ -1,59 +1,56 @@
 module t
 
-// +heroku goVersion go1.17
-go 1.17
+go 1.19
+
+//replace ariga.io/ogent => ../../
 
 require (
-	entgo.io/ent v0.10.0
-	github.com/go-chi/chi/v5 v5.0.4
-	github.com/jackc/pgx/v4 v4.15.0
-	github.com/lib/pq v1.10.4
-	github.com/mailru/easyjson v0.7.7
-	go.uber.org/zap v1.21.0
+	entgo.io/ent v0.11.9
+	github.com/go-faster/errors v0.6.1
+	github.com/go-faster/jx v0.42.0-alpha.1
+	github.com/mattn/go-sqlite3 v1.14.16
+	github.com/mazen160/go-random v0.0.0-20210308102632-d2b501c85c03
+	github.com/ogen-go/ogen v0.59.0
+	go.opentelemetry.io/otel v1.13.0
+	go.opentelemetry.io/otel/metric v0.36.0
+	go.opentelemetry.io/otel/trace v1.13.0
+	go.uber.org/multierr v1.9.0
 )
 
 require (
-	entgo.io/contrib v0.2.1-0.20220210075301-2b75bf138815 // indirect
+	ariga.io/atlas v0.9.1 // indirect
+	ariga.io/entviz v0.0.0-20230125130633-6c9be8e08c7c // indirect
+	ariga.io/ogent v0.0.0-20230309073626-8dc564a6a73e // indirect
+	entgo.io/contrib v0.3.5 // indirect
+	github.com/agext/levenshtein v1.2.3 // indirect
 	github.com/apparentlymart/go-textseg/v13 v13.0.0 // indirect
-	github.com/ariga/ogent v0.0.0-20220224071349-f1dbffa2e32a // indirect
+	github.com/dlclark/regexp2 v1.8.0 // indirect
+	github.com/fatih/color v1.14.1 // indirect
 	github.com/ghodss/yaml v1.0.0 // indirect
-	github.com/go-logr/logr v1.2.2 // indirect
+	github.com/go-faster/yamlx v0.4.1 // indirect
+	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-openapi/inflect v0.19.0 // indirect
 	github.com/golang/mock v1.6.0 // indirect
-	github.com/google/go-cmp v0.5.7 // indirect
-	github.com/jackc/chunkreader/v2 v2.0.1 // indirect
-	github.com/jackc/pgconn v1.11.0 // indirect
-	github.com/jackc/pgio v1.0.0 // indirect
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgproto3/v2 v2.2.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20200714003250-2b9c44734f2b // indirect
-	github.com/jackc/pgtype v1.10.0 // indirect
-	github.com/josharian/intern v1.0.0 // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/hashicorp/hcl/v2 v2.15.0 // indirect
 	github.com/kyokomi/lottery v1.2.0 // indirect
-	github.com/segmentio/asm v1.1.3 // indirect
-	github.com/uniplaces/carbon v0.1.6 // indirect
-	golang.org/x/crypto v0.0.0-20220112180741-5e0467b6c7ce // indirect
-	golang.org/x/mod v0.5.1 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
-
-require (
-	ariga.io/atlas v0.3.5 // indirect
-	github.com/agext/levenshtein v1.2.3 // indirect
-	github.com/go-faster/errors v0.5.0
-	github.com/go-faster/jx v0.32.2
-	github.com/google/uuid v1.3.0
-	github.com/hashicorp/hcl/v2 v2.11.1 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
 	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
-	github.com/ogen-go/ogen v0.19.0
+	github.com/segmentio/asm v1.2.0 // indirect
 	github.com/sergi/go-diff v1.1.0 // indirect
-	github.com/zclconf/go-cty v1.10.0 // indirect
-	go.opentelemetry.io/otel v1.4.1
-	go.opentelemetry.io/otel/metric v0.27.0
-	go.opentelemetry.io/otel/trace v1.4.1
-	go.uber.org/atomic v1.9.0 // indirect
-	go.uber.org/multierr v1.7.0 // indirect
-	golang.org/x/sys v0.0.0-20220222200937-f2425489ef4c // indirect
+	github.com/stoewer/go-strcase v1.2.0 // indirect
+	github.com/zclconf/go-cty v1.12.1 // indirect
+	go.uber.org/atomic v1.10.0 // indirect
+	go.uber.org/zap v1.24.0 // indirect
+	golang.org/x/exp v0.0.0-20230206171751-46f607a40771 // indirect
+	golang.org/x/mod v0.8.0 // indirect
+	golang.org/x/net v0.7.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	golang.org/x/text v0.7.0 // indirect
+	golang.org/x/tools v0.6.1-0.20230222164832-25d2519c8696 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
