@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package migrate
 
@@ -8,39 +8,87 @@ import (
 )
 
 var (
+	// CardsColumns holds the columns for the "cards" table.
+	CardsColumns = []*schema.Column{
+		{Name: "id", Type: field.TypeInt, Increment: true},
+		{Name: "card", Type: field.TypeInt, Nullable: true},
+		{Name: "status", Type: field.TypeString, Nullable: true},
+		{Name: "cp", Type: field.TypeInt, Nullable: true},
+		{Name: "url", Type: field.TypeString, Nullable: true, Default: "https://card.syui.ai"},
+		{Name: "created_at", Type: field.TypeTime, Nullable: true},
+		{Name: "user_card", Type: field.TypeInt},
+	}
+	// CardsTable holds the schema information for the "cards" table.
+	CardsTable = &schema.Table{
+		Name:       "cards",
+		Columns:    CardsColumns,
+		PrimaryKey: []*schema.Column{CardsColumns[0]},
+		ForeignKeys: []*schema.ForeignKey{
+			{
+				Symbol:     "cards_users_card",
+				Columns:    []*schema.Column{CardsColumns[6]},
+				RefColumns: []*schema.Column{UsersColumns[0]},
+				OnDelete:   schema.NoAction,
+			},
+		},
+	}
+	// GroupsColumns holds the columns for the "groups" table.
+	GroupsColumns = []*schema.Column{
+		{Name: "id", Type: field.TypeInt, Increment: true},
+		{Name: "name", Type: field.TypeString},
+	}
+	// GroupsTable holds the schema information for the "groups" table.
+	GroupsTable = &schema.Table{
+		Name:       "groups",
+		Columns:    GroupsColumns,
+		PrimaryKey: []*schema.Column{GroupsColumns[0]},
+		Indexes: []*schema.Index{
+			{
+				Name:    "group_name",
+				Unique:  true,
+				Columns: []*schema.Column{GroupsColumns[1]},
+			},
+		},
+	}
 	// UsersColumns holds the columns for the "users" table.
 	UsersColumns = []*schema.Column{
 		{Name: "id", Type: field.TypeInt, Increment: true},
-		{Name: "user", Type: field.TypeString, Unique: true, Size: 7},
-		{Name: "chara", Type: field.TypeString, Nullable: true, Default: "ponta"},
-		{Name: "skill", Type: field.TypeInt, Nullable: true, Default: 7},
-		{Name: "hp", Type: field.TypeInt, Nullable: true, Default: 7},
-		{Name: "attack", Type: field.TypeInt, Nullable: true, Default: 8},
-		{Name: "defense", Type: field.TypeInt, Nullable: true, Default: 19},
-		{Name: "critical", Type: field.TypeInt, Nullable: true, Default: 7},
-		{Name: "battle", Type: field.TypeInt, Nullable: true, Default: 1},
-		{Name: "win", Type: field.TypeInt, Nullable: true, Default: 0},
-		{Name: "day", Type: field.TypeInt, Nullable: true, Default: 0},
-		{Name: "percentage", Type: field.TypeFloat64, Nullable: true, Default: 0},
-		{Name: "limit", Type: field.TypeBool, Nullable: true, Default: false},
-		{Name: "status", Type: field.TypeString, Nullable: true, Default: "normal"},
-		{Name: "comment", Type: field.TypeString, Nullable: true, Default: ""},
+		{Name: "username", Type: field.TypeString, Unique: true, Size: 30},
 		{Name: "created_at", Type: field.TypeTime, Nullable: true},
-		{Name: "next", Type: field.TypeString, Nullable: true, Default: "20220617"},
 		{Name: "updated_at", Type: field.TypeTime, Nullable: true},
-		{Name: "url", Type: field.TypeString, Nullable: true, Default: "https://syui.cf/api"},
+		{Name: "next", Type: field.TypeString, Nullable: true, Default: "20230404"},
+		{Name: "group_users", Type: field.TypeInt, Nullable: true},
 	}
 	// UsersTable holds the schema information for the "users" table.
 	UsersTable = &schema.Table{
 		Name:       "users",
 		Columns:    UsersColumns,
 		PrimaryKey: []*schema.Column{UsersColumns[0]},
+		ForeignKeys: []*schema.ForeignKey{
+			{
+				Symbol:     "users_groups_users",
+				Columns:    []*schema.Column{UsersColumns[5]},
+				RefColumns: []*schema.Column{GroupsColumns[0]},
+				OnDelete:   schema.SetNull,
+			},
+		},
+		Indexes: []*schema.Index{
+			{
+				Name:    "user_username",
+				Unique:  true,
+				Columns: []*schema.Column{UsersColumns[1]},
+			},
+		},
 	}
 	// Tables holds all the tables in the schema.
 	Tables = []*schema.Table{
+		CardsTable,
+		GroupsTable,
 		UsersTable,
 	}
 )
 
 func init() {
+	CardsTable.ForeignKeys[0].RefTable = UsersTable
+	UsersTable.ForeignKeys[0].RefTable = GroupsTable
 }
