@@ -1,22 +1,74 @@
 // Code generated by ogen, DO NOT EDIT.
 package ogent
 
-type CreateUsersRes interface {
-	createUsersRes()
+type CreateCardRes interface {
+	createCardRes()
 }
 
-type DeleteUsersRes interface {
-	deleteUsersRes()
+type CreateGroupRes interface {
+	createGroupRes()
 }
 
-type ListUsersRes interface {
-	listUsersRes()
+type CreateUserRes interface {
+	createUserRes()
 }
 
-type ReadUsersRes interface {
-	readUsersRes()
+type DeleteCardRes interface {
+	deleteCardRes()
 }
 
-type UpdateUsersRes interface {
-	updateUsersRes()
+type DeleteGroupRes interface {
+	deleteGroupRes()
+}
+
+type DeleteUserRes interface {
+	deleteUserRes()
+}
+
+type ListCardRes interface {
+	listCardRes()
+}
+
+type ListGroupRes interface {
+	listGroupRes()
+}
+
+type ListGroupUsersRes interface {
+	listGroupUsersRes()
+}
+
+type ListUserCardRes interface {
+	listUserCardRes()
+}
+
+type ListUserRes interface {
+	listUserRes()
+}
+
+type ReadCardOwnerRes interface {
+	readCardOwnerRes()
+}
+
+type ReadCardRes interface {
+	readCardRes()
+}
+
+type ReadGroupRes interface {
+	readGroupRes()
+}
+
+type ReadUserRes interface {
+	readUserRes()
+}
+
+type UpdateCardRes interface {
+	updateCardRes()
+}
+
+type UpdateGroupRes interface {
+	updateGroupRes()
+}
+
+type UpdateUserRes interface {
+	updateUserRes()
 }
