@@ -3,250 +3,36 @@
 package ogent
 
 import (
-	"bytes"
-	"context"
-	"fmt"
-	"io"
-	"math"
-	"math/big"
-	"math/bits"
-	"net"
-	"net/http"
-	"net/url"
-	"regexp"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
 	"github.com/go-faster/errors"
-	"github.com/go-faster/jx"
-	"github.com/google/uuid"
-	"github.com/ogen-go/ogen/conv"
-	ht "github.com/ogen-go/ogen/http"
-	"github.com/ogen-go/ogen/json"
-	"github.com/ogen-go/ogen/otelogen"
-	"github.com/ogen-go/ogen/uri"
-	"github.com/ogen-go/ogen/validate"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/metric"
-	"go.opentelemetry.io/otel/trace"
 )
 
-// No-op definition for keeping imports.
-var (
-	_ = context.Background()
-	_ = fmt.Stringer(nil)
-	_ = strings.Builder{}
-	_ = errors.Is
-	_ = sort.Ints
-	_ = http.MethodGet
-	_ = io.Copy
-	_ = json.Marshal
-	_ = bytes.NewReader
-	_ = strconv.ParseInt
-	_ = time.Time{}
-	_ = conv.ToInt32
-	_ = uuid.UUID{}
-	_ = uri.PathEncoder{}
-	_ = url.URL{}
-	_ = math.Mod
-	_ = bits.LeadingZeros64
-	_ = big.Rat{}
-	_ = validate.Int{}
-	_ = ht.NewRequest
-	_ = net.IP{}
-	_ = otelogen.Version
-	_ = attribute.KeyValue{}
-	_ = trace.TraceIDFromHex
-	_ = otel.GetTracerProvider
-	_ = metric.NewNoopMeterProvider
-	_ = regexp.MustCompile
-	_ = jx.Null
-	_ = sync.Pool{}
-	_ = codes.Unset
-)
-
-func (s CreateUsersReq) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
-	}
-	return nil
-}
-func (s ListUsersOKApplicationJSON) Validate() error {
+func (s ListCardOKApplicationJSON) Validate() error {
 	if s == nil {
 		return errors.New("nil is invalid value")
 	}
-	var failures []validate.FieldError
-	for i, elem := range s {
-		if err := func() error {
-			if err := elem.Validate(); err != nil {
-				return err
-			}
-			return nil
-		}(); err != nil {
-			failures = append(failures, validate.FieldError{
-				Name:  fmt.Sprintf("[%d]", i),
-				Error: err,
-			})
-		}
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
+	return nil
+}
+func (s ListGroupOKApplicationJSON) Validate() error {
+	if s == nil {
+		return errors.New("nil is invalid value")
 	}
 	return nil
 }
-
-func (s UpdateUsersReq) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
+func (s ListGroupUsersOKApplicationJSON) Validate() error {
+	if s == nil {
+		return errors.New("nil is invalid value")
 	}
 	return nil
 }
-func (s UsersCreate) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
+func (s ListUserCardOKApplicationJSON) Validate() error {
+	if s == nil {
+		return errors.New("nil is invalid value")
 	}
 	return nil
 }
-func (s UsersList) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
-	}
-	return nil
-}
-func (s UsersRead) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
-	}
-	return nil
-}
-func (s UsersUpdate) Validate() error {
-	var failures []validate.FieldError
-	if err := func() error {
-		if s.Percentage.Set {
-			if err := func() error {
-				if err := (validate.Float{}).Validate(float64(s.Percentage.Value)); err != nil {
-					return errors.Wrap(err, "float")
-				}
-				return nil
-			}(); err != nil {
-				return err
-			}
-		}
-		return nil
-		return nil
-	}(); err != nil {
-		failures = append(failures, validate.FieldError{
-			Name:  "percentage",
-			Error: err,
-		})
-	}
-	if len(failures) > 0 {
-		return &validate.Error{Fields: failures}
+func (s ListUserOKApplicationJSON) Validate() error {
+	if s == nil {
+		return errors.New("nil is invalid value")
 	}
 	return nil
 }
