@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ogent
 
@@ -7,7 +7,9 @@ import (
 	"net/http"
 
 	"t/ent"
-	"t/ent/users"
+	"t/ent/card"
+	"t/ent/group"
+	"t/ent/user"
 
 	"github.com/go-faster/jx"
 )
@@ -27,63 +29,27 @@ func rawError(err error) jx.Raw {
 	return e.Bytes()
 }
 
-// CreateUsers handles POST /users-slice requests.
-func (h *OgentHandler) CreateUsers(ctx context.Context, req CreateUsersReq) (CreateUsersRes, error) {
-	b := h.client.Users.Create()
+// CreateCard handles POST /cards requests.
+func (h *OgentHandler) CreateCard(ctx context.Context, req *CreateCardReq) (CreateCardRes, error) {
+	b := h.client.Card.Create()
 	// Add all fields.
-	b.SetUser(req.User)
-	if v, ok := req.Chara.Get(); ok {
-		b.SetChara(v)
-	}
-	if v, ok := req.Skill.Get(); ok {
-		b.SetSkill(v)
-	}
-	if v, ok := req.Hp.Get(); ok {
-		b.SetHp(v)
-	}
-	if v, ok := req.Attack.Get(); ok {
-		b.SetAttack(v)
-	}
-	if v, ok := req.Defense.Get(); ok {
-		b.SetDefense(v)
-	}
-	if v, ok := req.Critical.Get(); ok {
-		b.SetCritical(v)
-	}
-	if v, ok := req.Battle.Get(); ok {
-		b.SetBattle(v)
-	}
-	if v, ok := req.Win.Get(); ok {
-		b.SetWin(v)
-	}
-	if v, ok := req.Day.Get(); ok {
-		b.SetDay(v)
-	}
-	if v, ok := req.Percentage.Get(); ok {
-		b.SetPercentage(v)
-	}
-	if v, ok := req.Limit.Get(); ok {
-		b.SetLimit(v)
+	if v, ok := req.Card.Get(); ok {
+		b.SetCard(v)
 	}
 	if v, ok := req.Status.Get(); ok {
 		b.SetStatus(v)
 	}
-	if v, ok := req.Comment.Get(); ok {
-		b.SetComment(v)
-	}
-	if v, ok := req.CreatedAt.Get(); ok {
-		b.SetCreatedAt(v)
-	}
-	if v, ok := req.Next.Get(); ok {
-		b.SetNext(v)
-	}
-	if v, ok := req.UpdatedAt.Get(); ok {
-		b.SetUpdatedAt(v)
+	if v, ok := req.Cp.Get(); ok {
+		b.SetCp(v)
 	}
 	if v, ok := req.URL.Get(); ok {
 		b.SetURL(v)
 	}
+	if v, ok := req.CreatedAt.Get(); ok {
+		b.SetCreatedAt(v)
+	}
 	// Add all edges.
+	b.SetOwnerID(req.Owner)
 	// Persist to storage.
 	e, err := b.Save(ctx)
 	if err != nil {
@@ -106,18 +72,18 @@ func (h *OgentHandler) CreateUsers(ctx context.Context, req CreateUsersReq) (Cre
 		}
 	}
 	// Reload the entity to attach all eager-loaded edges.
-	q := h.client.Users.Query().Where(users.ID(e.ID))
+	q := h.client.Card.Query().Where(card.ID(e.ID))
 	e, err = q.Only(ctx)
 	if err != nil {
 		// This should never happen.
 		return nil, err
 	}
-	return NewUsersCreate(e), nil
+	return NewCardCreate(e), nil
 }
 
-// ReadUsers handles GET /users-slice/{id} requests.
-func (h *OgentHandler) ReadUsers(ctx context.Context, params ReadUsersParams) (ReadUsersRes, error) {
-	q := h.client.Users.Query().Where(users.IDEQ(params.ID))
+// ReadCard handles GET /cards/{id} requests.
+func (h *OgentHandler) ReadCard(ctx context.Context, params ReadCardParams) (ReadCardRes, error) {
+	q := h.client.Card.Query().Where(card.IDEQ(params.ID))
 	e, err := q.Only(ctx)
 	if err != nil {
 		switch {
@@ -138,50 +104,17 @@ func (h *OgentHandler) ReadUsers(ctx context.Context, params ReadUsersParams) (R
 			return nil, err
 		}
 	}
-	return NewUsersRead(e), nil
+	return NewCardRead(e), nil
 }
 
-// UpdateUsers handles PATCH /users-slice/{id} requests.
-func (h *OgentHandler) UpdateUsers(ctx context.Context, req UpdateUsersReq, params UpdateUsersParams) (UpdateUsersRes, error) {
-	b := h.client.Users.UpdateOneID(params.ID)
+// UpdateCard handles PATCH /cards/{id} requests.
+func (h *OgentHandler) UpdateCard(ctx context.Context, req *UpdateCardReq, params UpdateCardParams) (UpdateCardRes, error) {
+	b := h.client.Card.UpdateOneID(params.ID)
 	// Add all fields.
-	if v, ok := req.Hp.Get(); ok {
-		b.SetHp(v)
-	}
-	if v, ok := req.Attack.Get(); ok {
-		b.SetAttack(v)
-	}
-	if v, ok := req.Defense.Get(); ok {
-		b.SetDefense(v)
-	}
-	if v, ok := req.Critical.Get(); ok {
-		b.SetCritical(v)
-	}
-	if v, ok := req.Battle.Get(); ok {
-		b.SetBattle(v)
-	}
-	if v, ok := req.Win.Get(); ok {
-		b.SetWin(v)
-	}
-	if v, ok := req.Day.Get(); ok {
-		b.SetDay(v)
-	}
-	if v, ok := req.Percentage.Get(); ok {
-		b.SetPercentage(v)
-	}
-	if v, ok := req.Limit.Get(); ok {
-		b.SetLimit(v)
-	}
-	if v, ok := req.Comment.Get(); ok {
-		b.SetComment(v)
-	}
-	if v, ok := req.Next.Get(); ok {
-		b.SetNext(v)
-	}
-	if v, ok := req.UpdatedAt.Get(); ok {
-		b.SetUpdatedAt(v)
-	}
 	// Add all edges.
+	if v, ok := req.Owner.Get(); ok {
+		b.SetOwnerID(v)
+	}
 	// Persist to storage.
 	e, err := b.Save(ctx)
 	if err != nil {
@@ -204,18 +137,18 @@ func (h *OgentHandler) UpdateUsers(ctx context.Context, req UpdateUsersReq, para
 		}
 	}
 	// Reload the entity to attach all eager-loaded edges.
-	q := h.client.Users.Query().Where(users.ID(e.ID))
+	q := h.client.Card.Query().Where(card.ID(e.ID))
 	e, err = q.Only(ctx)
 	if err != nil {
 		// This should never happen.
 		return nil, err
 	}
-	return NewUsersUpdate(e), nil
+	return NewCardUpdate(e), nil
 }
 
-// DeleteUsers handles DELETE /users-slice/{id} requests.
-func (h *OgentHandler) DeleteUsers(ctx context.Context, params DeleteUsersParams) (DeleteUsersRes, error) {
-	err := h.client.Users.DeleteOneID(params.ID).Exec(ctx)
+// DeleteCard handles DELETE /cards/{id} requests.
+func (h *OgentHandler) DeleteCard(ctx context.Context, params DeleteCardParams) (DeleteCardRes, error) {
+	err := h.client.Card.DeleteOneID(params.ID).Exec(ctx)
 	if err != nil {
 		switch {
 		case ent.IsNotFound(err):
@@ -235,13 +168,13 @@ func (h *OgentHandler) DeleteUsers(ctx context.Context, params DeleteUsersParams
 			return nil, err
 		}
 	}
-	return new(DeleteUsersNoContent), nil
+	return new(DeleteCardNoContent), nil
 
 }
 
-// ListUsers handles GET /users-slice requests.
-func (h *OgentHandler) ListUsers(ctx context.Context, params ListUsersParams) (ListUsersRes, error) {
-	q := h.client.Users.Query()
+// ListCard handles GET /cards requests.
+func (h *OgentHandler) ListCard(ctx context.Context, params ListCardParams) (ListCardRes, error) {
+	q := h.client.Card.Query()
 	page := 1
 	if v, ok := params.Page.Get(); ok {
 		page = v
@@ -250,7 +183,9 @@ func (h *OgentHandler) ListUsers(ctx context.Context, params ListUsersParams) (L
 	if v, ok := params.ItemsPerPage.Get(); ok {
 		itemsPerPage = v
 	}
-	es, err := q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage).All(ctx)
+	q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage)
+
+	es, err := q.All(ctx)
 	if err != nil {
 		switch {
 		case ent.IsNotFound(err):
@@ -270,5 +205,454 @@ func (h *OgentHandler) ListUsers(ctx context.Context, params ListUsersParams) (L
 			return nil, err
 		}
 	}
-	return ListUsersOKApplicationJSON(NewUsersLists(es)), nil
+	r := NewCardLists(es)
+	return (*ListCardOKApplicationJSON)(&r), nil
+}
+
+// ReadCardOwner handles GET /cards/{id}/owner requests.
+func (h *OgentHandler) ReadCardOwner(ctx context.Context, params ReadCardOwnerParams) (ReadCardOwnerRes, error) {
+	q := h.client.Card.Query().Where(card.IDEQ(params.ID)).QueryOwner()
+	e, err := q.Only(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	return NewCardOwnerRead(e), nil
+}
+
+// CreateGroup handles POST /groups requests.
+func (h *OgentHandler) CreateGroup(ctx context.Context, req *CreateGroupReq) (CreateGroupRes, error) {
+	b := h.client.Group.Create()
+	// Add all fields.
+	b.SetName(req.Name)
+	// Add all edges.
+	b.AddUserIDs(req.Users...)
+	// Persist to storage.
+	e, err := b.Save(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	// Reload the entity to attach all eager-loaded edges.
+	q := h.client.Group.Query().Where(group.ID(e.ID))
+	e, err = q.Only(ctx)
+	if err != nil {
+		// This should never happen.
+		return nil, err
+	}
+	return NewGroupCreate(e), nil
+}
+
+// ReadGroup handles GET /groups/{id} requests.
+func (h *OgentHandler) ReadGroup(ctx context.Context, params ReadGroupParams) (ReadGroupRes, error) {
+	q := h.client.Group.Query().Where(group.IDEQ(params.ID))
+	e, err := q.Only(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	return NewGroupRead(e), nil
+}
+
+// UpdateGroup handles PATCH /groups/{id} requests.
+func (h *OgentHandler) UpdateGroup(ctx context.Context, req *UpdateGroupReq, params UpdateGroupParams) (UpdateGroupRes, error) {
+	b := h.client.Group.UpdateOneID(params.ID)
+	// Add all fields.
+	if v, ok := req.Name.Get(); ok {
+		b.SetName(v)
+	}
+	// Add all edges.
+	if req.Users != nil {
+		b.ClearUsers().AddUserIDs(req.Users...)
+	}
+	// Persist to storage.
+	e, err := b.Save(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	// Reload the entity to attach all eager-loaded edges.
+	q := h.client.Group.Query().Where(group.ID(e.ID))
+	e, err = q.Only(ctx)
+	if err != nil {
+		// This should never happen.
+		return nil, err
+	}
+	return NewGroupUpdate(e), nil
+}
+
+// DeleteGroup handles DELETE /groups/{id} requests.
+func (h *OgentHandler) DeleteGroup(ctx context.Context, params DeleteGroupParams) (DeleteGroupRes, error) {
+	err := h.client.Group.DeleteOneID(params.ID).Exec(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	return new(DeleteGroupNoContent), nil
+
+}
+
+// ListGroup handles GET /groups requests.
+func (h *OgentHandler) ListGroup(ctx context.Context, params ListGroupParams) (ListGroupRes, error) {
+	q := h.client.Group.Query()
+	page := 1
+	if v, ok := params.Page.Get(); ok {
+		page = v
+	}
+	itemsPerPage := 30
+	if v, ok := params.ItemsPerPage.Get(); ok {
+		itemsPerPage = v
+	}
+	q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage)
+
+	es, err := q.All(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	r := NewGroupLists(es)
+	return (*ListGroupOKApplicationJSON)(&r), nil
+}
+
+// ListGroupUsers handles GET /groups/{id}/users requests.
+func (h *OgentHandler) ListGroupUsers(ctx context.Context, params ListGroupUsersParams) (ListGroupUsersRes, error) {
+	q := h.client.Group.Query().Where(group.IDEQ(params.ID)).QueryUsers()
+	page := 1
+	if v, ok := params.Page.Get(); ok {
+		page = v
+	}
+	itemsPerPage := 30
+	if v, ok := params.ItemsPerPage.Get(); ok {
+		itemsPerPage = v
+	}
+	q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage)
+	es, err := q.All(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	r := NewGroupUsersLists(es)
+	return (*ListGroupUsersOKApplicationJSON)(&r), nil
+}
+
+// CreateUser handles POST /users requests.
+func (h *OgentHandler) CreateUser(ctx context.Context, req *CreateUserReq) (CreateUserRes, error) {
+	b := h.client.User.Create()
+	// Add all fields.
+	b.SetUsername(req.Username)
+	if v, ok := req.CreatedAt.Get(); ok {
+		b.SetCreatedAt(v)
+	}
+	if v, ok := req.UpdatedAt.Get(); ok {
+		b.SetUpdatedAt(v)
+	}
+	if v, ok := req.Next.Get(); ok {
+		b.SetNext(v)
+	}
+	// Add all edges.
+	b.AddCardIDs(req.Card...)
+	// Persist to storage.
+	e, err := b.Save(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	// Reload the entity to attach all eager-loaded edges.
+	q := h.client.User.Query().Where(user.ID(e.ID))
+	e, err = q.Only(ctx)
+	if err != nil {
+		// This should never happen.
+		return nil, err
+	}
+	return NewUserCreate(e), nil
+}
+
+// ReadUser handles GET /users/{id} requests.
+func (h *OgentHandler) ReadUser(ctx context.Context, params ReadUserParams) (ReadUserRes, error) {
+	q := h.client.User.Query().Where(user.IDEQ(params.ID))
+	e, err := q.Only(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	return NewUserRead(e), nil
+}
+
+// UpdateUser handles PATCH /users/{id} requests.
+func (h *OgentHandler) UpdateUser(ctx context.Context, req *UpdateUserReq, params UpdateUserParams) (UpdateUserRes, error) {
+	b := h.client.User.UpdateOneID(params.ID)
+	// Add all fields.
+	if v, ok := req.UpdatedAt.Get(); ok {
+		b.SetUpdatedAt(v)
+	}
+	if v, ok := req.Next.Get(); ok {
+		b.SetNext(v)
+	}
+	// Add all edges.
+	if req.Card != nil {
+		b.ClearCard().AddCardIDs(req.Card...)
+	}
+	// Persist to storage.
+	e, err := b.Save(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	// Reload the entity to attach all eager-loaded edges.
+	q := h.client.User.Query().Where(user.ID(e.ID))
+	e, err = q.Only(ctx)
+	if err != nil {
+		// This should never happen.
+		return nil, err
+	}
+	return NewUserUpdate(e), nil
+}
+
+// DeleteUser handles DELETE /users/{id} requests.
+func (h *OgentHandler) DeleteUser(ctx context.Context, params DeleteUserParams) (DeleteUserRes, error) {
+	err := h.client.User.DeleteOneID(params.ID).Exec(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsConstraintError(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	return new(DeleteUserNoContent), nil
+
+}
+
+// ListUser handles GET /users requests.
+func (h *OgentHandler) ListUser(ctx context.Context, params ListUserParams) (ListUserRes, error) {
+	q := h.client.User.Query()
+	page := 1
+	if v, ok := params.Page.Get(); ok {
+		page = v
+	}
+	itemsPerPage := 30
+	if v, ok := params.ItemsPerPage.Get(); ok {
+		itemsPerPage = v
+	}
+	q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage)
+
+	es, err := q.All(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	r := NewUserLists(es)
+	return (*ListUserOKApplicationJSON)(&r), nil
+}
+
+// ListUserCard handles GET /users/{id}/card requests.
+func (h *OgentHandler) ListUserCard(ctx context.Context, params ListUserCardParams) (ListUserCardRes, error) {
+	q := h.client.User.Query().Where(user.IDEQ(params.ID)).QueryCard()
+	page := 1
+	if v, ok := params.Page.Get(); ok {
+		page = v
+	}
+	itemsPerPage := 30
+	if v, ok := params.ItemsPerPage.Get(); ok {
+		itemsPerPage = v
+	}
+	q.Limit(itemsPerPage).Offset((page - 1) * itemsPerPage)
+	es, err := q.All(ctx)
+	if err != nil {
+		switch {
+		case ent.IsNotFound(err):
+			return &R404{
+				Code:   http.StatusNotFound,
+				Status: http.StatusText(http.StatusNotFound),
+				Errors: rawError(err),
+			}, nil
+		case ent.IsNotSingular(err):
+			return &R409{
+				Code:   http.StatusConflict,
+				Status: http.StatusText(http.StatusConflict),
+				Errors: rawError(err),
+			}, nil
+		default:
+			// Let the server handle the error.
+			return nil, err
+		}
+	}
+	r := NewUserCardLists(es)
+	return (*ListUserCardOKApplicationJSON)(&r), nil
 }
