@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package predicate
 
@@ -6,5 +6,11 @@ import (
 	"entgo.io/ent/dialect/sql"
 )
 
-// Users is the predicate function for users builders.
-type Users func(*sql.Selector)
+// Card is the predicate function for card builders.
+type Card func(*sql.Selector)
+
+// Group is the predicate function for group builders.
+type Group func(*sql.Selector)
+
+// User is the predicate function for user builders.
+type User func(*sql.Selector)
